Rename icons folder to svg
