+++ src/assets/svg/index.tsx
