+++ src/components/AerosolveLogo/AerosolveLogo.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
-import { ReactComponent as AL } from 'assets/images/AerosolveLogo.svg';
+import React from 'react'
+import { ReactComponent as AL } from 'assets/svg/AerosolveLogo.svg'
 
 export const AerosolveLogo: React.FC<{}> = (props) => {
-  return <AL width="100%" height="100%" viewBox="0 0 230 91" />;
-};
+  return <AL width="100%" height="100%" viewBox="0 0 230 91" />
+}
